ImagePicker: Adds loading, loaded and error icons to upload previews.
This commit adds loading, loaded and error icons to upload image
preview thumbnails. This means that if when an image is uploading
you will see a spinner, checkmark when its done and warning sign if
theres an error.

This commit also takes special care to make sure that if connectivity
is dropped for some reason while uploading images on slower connections,
you can pick up where you left off.

We also take care of the case that the connection is dropped but it
comes back, the user clicks the upload button again, and the old api
call that now times out should not show an error when the user has
already started a fresh upload attempt.