ComposeBox: Allow selecting multiple images per message with previews.
This commit makes the necessary changes to ComposeMenu and ComposeBox
in order to select and upload multiple images per message while composing.
It also enables showing thumbnail previews of images selected for upload.

This is done by editing ComposeMenu.js to allow selecting multiple images
and by editing ComposeBox on iOS and android to be able to handle these
selections and thereafter upload them on send.

It also adds some styling to be able to display the image preview thumbnails
correctly.