ImagePicker: Change ImagePicker Library Implementation.
Update the implementation of image uploading in ComposeMenu.js
We adopt react-native-image-crop-picker in favor of the
old library (react-native-image-picker).

This gives us the additional ability to downsample and downscale images
before uploading them and also allows selecting multiple images in the
same selection ( iOS only for now ). These changes will be implemented in
a following commit.

Currently, there is a bug in the image picker which results in the max
width and height to only be applied on android. This is still an
improvement over the previous library but will need to be kept in mind
so when that bug is fixed we can upgrade.
See https://github.com/ivpusic/react-native-image-crop-picker/issues/604